try to solve format problem at iphone

+++ src/process.rs
@@ -4,7 +4,8 @@ use crate::upload;
 use crate::youtube::download_video;
 
 pub(crate) async fn youtube(
-    url: &str, platform: &str, output: &str, rutube_api_key: Option<String>,
+    url: &str, platform: &str, output: &str,
+    delete_youtube: bool, delete_transformed: bool, rutube_api_key: Option<String>,
     bot_api_url: &str, max_file_size: u64, bot_token: Option<String>,
     chat_id: Option<i64>, vk_access_token: Option<String>) -> anyhow::Result<()> {
 
@@ -21,8 +22,11 @@ pub(crate) async fn youtube(
     upload::upload(platform, &transformed_file, &title, rutube_api_key, &bot_api_url,
                    max_file_size, bot_token, chat_id, vk_access_token, url, "").await?;
 
-    fs::remove_file(&downloaded_file)?;
-    fs::remove_file(&transformed_file)?;
-
+    if delete_youtube {
+        fs::remove_file(&downloaded_file)?;
+    }
+    if delete_transformed {
+        fs::remove_file(&transformed_file)?;
+    }
     Ok(())
 }
